board: nxp: ls1021a-tsn: Apply downstream kernel patches on top of net-next
These are in the process of being submitted mainline. At the moment,
they are needed in order to perform 802.1AS synchronization efficiently
and correctly.

Most of the patches are on sja1105 PTP. However those are the least
important.

The critical patch is
0001-gianfar-Fix-TX-timestamping-with-stacked-DSA-and-PHY.patch.
Without it, there is a risk of apparent PTP message reordering after a
while.

Signed-off-by: Vladimir Oltean <olteanv@gmail.com>