package/linuxptp: Integrate strict checking patch series
This was sent for upstream review here:
https://sourceforge.net/p/linuxptp/mailman/message/36880648/

And I was hoping for quicker acceptance, but it looks like it's taking a
while. And since we still have some unknown issues with PTP (this time
only on some specific boards), it is useful to have the strict checking
patches applied, even if downstream, in the hope that they might be able
to catch more bugs.

Signed-off-by: Vladimir Oltean <olteanv@gmail.com>