package/linuxptp: Add a configuration file for phc2sys
Some boards need custom settings for PTP, such as OpenIL boards which
ship by default with the gPTP.cfg profile. When using ptp4l with 802.1AS
(therefore when transportSpecific is 1), the phc2sys service needs to
have the same value for transportSpecific, otherwise it won't be able to
communicate through the UNIX domain socket.

And since the OpenIL boards provide a rootfs overlay for
/etc/linuxptp.cfg, move the phc2sys configuration options to a new
/etc/linuxptp-system-clock.cfg file as well, so that it can be
overridden in the same manner.

The new config file is also more flexible than it was before: the PHC
can now be both a local synchronization master or slave clock (depending
on the ptp4l sync direction), it auto-adjusts to loss of sync quicker
via clock stepping, and it can also serve CLOCK_REALTIME into the PTP
timing domain.

Signed-off-by: Vladimir Oltean <olteanv@gmail.com>