board: nxp: ls1021a-tsn: Kernel bugfix for VLAN filtering
The switch ports on the LS1021A-TSN board did not recognize VLAN tags
because of incorrect EtherType settings. So all traffic would be
classified as untagged, and the PCP from the frames would be ignored and
therefore there would also be no traffic class separation, which would
break the 802.1Qbv demonstration.

Signed-off-by: Vladimir Oltean <olteanv@gmail.com>