board: nxp: ls1021a-tsn: Add kernel patch for SPI tracepoints
The previous patch for sja1105 driver timestamping was not sufficient.
It is telling us that the PTP clock jumps back and forward in time by
around 2^30 ticks (1 tick is 8 ns), which either means a bit flip on the
SPI interface, or a bug in the kernel somewhere.

To know which one it is, we need to print the RX data from the SPI
interface as close to hardware as possible.

Signed-off-by: Vladimir Oltean <olteanv@gmail.com>