board: nxp: ls1021a-tsn: Implement SPI bit flips workaround for PTP time
Apparently on some LS1021A-TSN boards, it is possible to get bit flips
on bits 31-28 when reading the 64-bit PTPCLKVAL register (the current
PTP time). This throws off the PTP synchronization irrecoverably.

Using the fsl_dspi tracepoints it was determined that these bit flips
actually originate from the wire. The reason is unknown at the moment.

The sja1105 switch has CRC32 protection for the static config, but not
for dynamic config commands (including access to the PTP registers). So
that can't be used to detect errors here.

Implement a sanity check mechanism in the driver for estimating a valid
range for the PTP clock, based on passage of time on the local CPU. When
the PTP clock appears to jump outside the plausible range, drop the
readout and try again, up to 10 times.

Signed-off-by: Vladimir Oltean <olteanv@gmail.com>