Get rid of conditional patches
The LS1028A-RDB needs some custom RCW changes. Move those to a global
patch directory for this board instead of polluting the RCW makefile.

The transapi module appears to not be using conditional patches for any
reason. Rename them to .patch and remove the manual application from the
makefile.

The kernel patches for ls1028a-rdb had to be moved from board/nxp/common
to board/nxp/ls1028ardb because we can't have 2 BR2_GLOBAL_PATCH_DIR
variables, and because the conditional patches for RCW need to be
applied only for LS1028A-RDB.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>